requirements.txt
Flask==1.1.2
python-dotenv==0.15.0
Flask-SQLAlchemy==2.4.4
marshmallow-sqlalchemy==0.24.2
psycopg2==2.8.6